requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Qdrant (required by mem0)
qdrant-client==1.12.1

# Mem0 - Memory management
mem0ai==0.1.75

# OpenAI (httpx 0.27.x required: 0.28+ breaks with openai proxies kwarg)
openai==1.55.0
httpx>=0.27.0,<0.28.0

# MongoDB (motor includes pymongo)
motor==3.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib==1.7.4
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1
orjson==3.10.12

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
